rag_applications/cat-facts-web3/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
web3==6.4.0
python-dotenv==1.0.0
ipfshttpclient==0.8.0a2
ollama==0.1.4
pydantic==1.10.7
